backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scicalc-pro-backend"
version = "1.0.0"
description = "SciCalc Pro - Backend FastAPI con arquitectura hexagonal"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "mpmath>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "freezegun>=1.4.0",
    "httpx>=0.27.0",
    "selenium>=4.18.0",
    "webdriver-manager>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml:coverage.xml"
markers = ["e2e: End-to-end Selenium tests (require Chrome + PostgreSQL via DATABASE_URL)"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "src/infrastructure/persistence/migrations/*"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
